% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Descriptor.R
\name{describe}
\alias{describe}
\alias{describe}
\alias{describe_if}
\title{Specify descriptors for a table}
\usage{
describe(dscr, with, append = FALSE, ...)

describe_if(dscr, .predicate, with, ...)
}
\arguments{
\item{with}{list of descriptors to be used with the selected variables.}

\item{...}{one or more unquoted expressions separated by commas passed to
\code{\link[dplyr]{dplyr}}'s \code{\link[dplyr]{select}}.
All descriptors specified in \code{with} are used for the selected variables.
Currently not used for \code{describe_if}.}

\item{.predicate}{A predicate function passed to \code{\link[dplyr]{select_if}}.}

\item{descr}{\code{\link{describr}} object; the specified descriptors will
be added for the selected variables in \code{dscr}.}

\item{APPEND}{[=FALSE] flag indicating whether previously defined descriptors
should be overwritten.}
}
\value{
An object of class \code{\link{describr}} where the defined descriptors
  are added to the specified variables.
}
\description{
\code{describe} can be used to specify \emph{which} variables should be
described \emph{how}.

\code{describe_if} allows to to specify variables by using predicates as in
\code{\link[dplyr]{select_if}}.
}
