% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describr.R
\docType{package}
\name{describr-package}
\alias{describr-package}
\title{describr: making univariate descriptive tables fun in R}
\description{
This package provides functionality to define descriptive tables
with minimum effort and human-readable code.
It is designed to work well with the tidyverse and heavily relies on
functionality of \code{\link[dplyr]{select}} and
\code{\link[dplyr]{select_if}} of the \code{\link[dplyr]{dplyr}} package.

The central \code{\link{describr}} object is the basis for defining a
table.
}
