% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describr.R
\name{describr}
\alias{describr}
\title{\code{describr} object}
\usage{
describr(df, by = NULL, theme_new = theme_default(), pvalues = FALSE,
  totals = TRUE)
}
\arguments{
\item{df}{a data frame.}

\item{by}{unquoted name of optional stratifying factor}

\item{theme_new}{theme to use}

\item{pvalus}{Flag indicating whether p-values for homogeneity null hypothesis
should be printed if available.}

\item{total}{Flag indicating whether all descriptors should also be given
for the total sample.}
}
\value{
An object of class \code{describr < list} with fields:
  \describe{
     \item{\code{df}}{data frame to describe}
     \item{\code{by}}{variable name by which to stratify the description
       (character or NULL)}
     \item{\code{core}}{named list with names corresponding
       to the variables in \code{df} which will b described and eac element
       is a list of \code{descriptor} objects osed to describe that variable.
       TODO: rename!}
     \item{\code{group_descriptors}}{list of descriptors used for the grouping
       (stratifying) variable \code{by}.}
     \item{\code{pvalues}}{Boolean flag indicating whether p-values should be
       displayed}
     \item{\code{totals}}{Boolean flag indicating whether a total column
       should be included when the description is stratified by a grouping
       factor.}
     \item{\code{theme_new}}{The theme to use for plotting. TODO: rename}
  }
}
\description{
This is the core object for \code{describr} package.
All information required to create a descriptive table for a data frame is
stored a \code{describr}-object.
}
