% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{dscr_anderson_darling}
\alias{dscr_anderson_darling}
\title{p-value for Anderson-Darling test (homogeineity of distributions)}
\usage{
dscr_anderson_darling(label = "Anderson-Darling", version = 1)
}
\description{
Computes asymptotic p-values for k-sample ANderson-Darling test using
\code{\link[kSamples]{ad.test}}.
}
