% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{dscr_chisq}
\alias{dscr_chisq}
\title{p-value for Chi-squared test (homogeineity of distributions)}
\usage{
dscr_chisq(label_one_sample = "Chi-squared test",
  label_cross_table = "Chi-squared test")
}
\description{
Computes p-values for Chi-squared test using \code{\link{chisq.test}}.
A one-sample Chi-squared test can be conducted by using \code{groupname==varname}.
}
