% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{dscr_cross_table_chisq}
\alias{dscr_cross_table_chisq}
\title{p-value for chi-squared test (homogeineity of distributions)}
\usage{
dscr_cross_table_chisq(label = "Chi-squared test")
}
\description{
Computes cross-table p-values for Chi-squared test using \code{\link{chisq.test}}.
}
