% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{dscr_Kruskal}
\alias{dscr_Kruskal}
\title{p-value for Kruskal-Wallis test (homogeineity of locations)}
\usage{
dscr_Kruskal(label_two_groups = "Mann-Whitney test",
  label_more_groups = "Kruskal-Wallis test")
}
\description{
Computes p-values for Wilcoxon rank-sum test (two groups) or Kruskal-Wallis
test (more levels) using \code{\link{wilcox.test}} and
\code{\link{kruskal.test}}.
Both tests are equivalent but an exact p-value is only returned for the
two-group case.
}
