% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{dscr_no_test}
\alias{dscr_no_test}
\title{Dummy class if a descriptor does not provide a p-value}
\usage{
dscr_no_test(label = "no test performed")
}
\description{
Dummy class if a descriptor does not provide a p-value
}
