% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{dscr_one_sample_chisq}
\alias{dscr_one_sample_chisq}
\title{p-value for chi-squared test (homogeineity of distributions)}
\usage{

  dscr_one_sample_chisq(label = "one-sample Chi-squared test for discrete uniformity")
}
\description{
Computes one-sample p-values for chi-squared test using \code{\link{chisq.test}}
(deviation from discrete uniform distribution).
}
