% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{dscr_Welch_ANOVA}
\alias{dscr_Welch_ANOVA}
\title{p-value for Welch test (homogeineity of means)}
\usage{
dscr_Welch_ANOVA(label_two_groups = "Welch test",
  label_more_groups = "Welch ANOVA")
}
\description{
Computes p-values for Welch test (two groups) or Welch ANOVA (more levels)
using \code{\link{oneway.test}}.#'
}
