% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{label}
\alias{label}
\title{Get a label string for the corresponding p-value}
\usage{
label(ht, df, varname, groupname, ...)
}
\arguments{
\item{df}{data frame}

\item{varname}{character string giving name of variable to test for homogeneity
with respect to group.}

\item{groupname}{character string giving name of grouping-variable (a factor)
in df.}

\item{p}{p-value object}
}
\value{
A string identifying the method used to compute the corresponding
  p-value.
}
\description{
Get a label string for the corresponding p-value
}
