% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describr.R
\name{make_plottable}
\alias{make_plottable}
\alias{make_plottable}
\alias{make_plottable.describr}
\title{Make object plottable using \code{grid}}
\usage{
make_plottable(x, ...)

\method{make_plottable}{describr}(dscr, ...)
}
\arguments{
\item{x}{object to plot.}
}
\description{
\code{make_plottable} returns an object which can be plotted using
\code{\link[grid]{grid.draw}}.
}
