% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_pdf.R
\name{to_pdf}
\alias{to_pdf}
\alias{to_pdf.describr}
\alias{to_pdf.describrGtable}
\alias{to_pdf.list}
\title{Create PDF file from plottable object}
\usage{
to_pdf(x, name, ...)

\method{to_pdf}{describr}(x, name, maxwidth = Inf, maxheight = Inf,
  ...)

\method{to_pdf}{describrGtable}(x, name, maxwidth = Inf,
  maxheight = Inf, ...)

\method{to_pdf}{list}(x, name, ...)
}
\arguments{
\item{x}{object to plot.}

\item{name}{name used for output file(s), might be extended if multiple plots
are generated and no ending is needed (e.g. 'my_file').}

\item{...}{further options for specific methods.}

\item{maxwidth}{maximal plotting width in inches (plain numeric).}

\item{maxheight}{maximal plotting heigth in inches (plain numeric).}
}
\description{
\code{to_pdf} plots an object directly into one ore multiple PDF files.
}
