% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_png.R
\name{to_png}
\alias{to_png}
\alias{to_png.describr}
\alias{to_png.describrGtable}
\alias{to_png.list}
\title{Create PNG file from plottable object}
\usage{
to_png(x, name, ...)

\method{to_png}{describr}(x, name, maxwidth = Inf, maxheight = Inf,
  dpi = 96, ...)

\method{to_png}{describrGtable}(x, name, maxwidth = Inf,
  maxheight = Inf, dpi = 96, ...)

\method{to_png}{list}(x, name, dpi = 96, ...)
}
\arguments{
\item{x}{object to plot.}

\item{name}{name used for output file(s), might be extended if multiple plots
are generated and no ending is needed (e.g. 'my_file').}

\item{...}{further options for specific methods.}

\item{maxwidth}{maximal plotting width in inches (plain numeric).}

\item{maxheight}{maximal plotting heigth in inches (plain numeric).}

\item{dpi}{dots per inch}
}
\description{
\code{to_pdf} plots an object directly into one ore multiple PDF files.
}
