% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvalues.R
\name{label}
\alias{label}
\title{Get a label string from object}
\usage{
label(x, ...)
}
\arguments{
\item{x}{object}

\item{...}{other parameters passed to methods}
}
\value{
A string identifying x.
}
\description{
Get a label string from object
}
